SB-24240 moved test-cases to another class
