+++ actors/sunbird-lms-mw/actors/common/src/test/java/org/sunbird/learner/util/FormApiUtilHandlerTest.java
@@ -36,26 +36,6 @@ public class FormApiUtilHandlerTest {
   }
 
   @Test
-  public void testGetFormApiEmptyConfig() {
-
-    when(HttpClientUtil.post(Mockito.anyString(), Mockito.anyString(), Mockito.anyObject()))
-        .thenReturn("");
-    Map<String, Object> dataConfigMap =
-        FormApiUtil.getProfileConfig("locationCode", new RequestContext());
-    Assert.assertEquals(0, dataConfigMap.size());
-  }
-
-  @Test
-  public void testGetFormApiEmptyResponseConfig() {
-
-    when(HttpClientUtil.post(Mockito.anyString(), Mockito.anyString(), Mockito.anyObject()))
-        .thenReturn(getFormApiEmptyResponse());
-    Map<String, Object> dataConfigMap =
-        FormApiUtil.getProfileConfig("locationCode", new RequestContext());
-    Assert.assertEquals(0, dataConfigMap.size());
-  }
-
-  @Test
   public void testprepareFormApiUtilPayload() {
     FormUtilRequest req = new FormUtilRequest();
     req.setComponent("component");
@@ -81,14 +61,6 @@ public class FormApiUtilHandlerTest {
             + " { \"form\": { \"type\": \"profileconfig\", \"subtype\": \"28\", \"action\": \"get\", \"component\": \"*\", "
             + "\"framework\": \"*\", \"created_on\": \"2021-01-12T06:12:51.284Z\", \"last_modified_on\": null, \"rootOrgId\": "
             + "\"*\" } }, \"ts\": \"2021-01-13T22:29:26.748Z\", \"ver\": \"1.0\" }";
-    return formData;
-  }
-
-  public String getFormApiEmptyResponse() {
-    String formData =
-        "{ \"id\": \"api.form.read\", \"params\": { \"resmsgid\": \"5ebb6cb5-07a0-4407-8013-b45043270d7a\", \"msgid\": \"3af660bf-fc92-4c93-acd1-36ad81cb8f35\", \"status\": \"successful\" }, \"responseCode\": \"OK\", "
-            + "\"result\":"
-            + ", \"ts\": \"2021-01-13T22:29:26.748Z\", \"ver\": \"1.0\" }";
     return formData;
   }
 }
